+++ src/realm/object-store/thread_safe_reference.cpp
@@ -167,7 +167,9 @@ public:
         : Payload(*r.get_realm())
         , m_ordering(r.get_descriptor_ordering())
     {
-        if (auto list = r.get_collection()) {
+        if (r.get_type() != PropertyType::Object) {
+            auto list = r.get_collection();
+            REALM_ASSERT(list);
             m_key = list->get_owner_key();
             m_table_key = list->get_table()->get_key();
             m_col_key = list->get_col_key();